refactor: move and isolate conversation context test
